requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
scikit-learn==1.3.1
PyPDF2==3.0.1
python-docx==1.0.1
python-dotenv==1.0.0
reportlab==4.0.7
